Task#5: move currentMovie logic in movieCard
