+++ client/src/core/store/rootReducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import { reducer as searchMoviesReducer } from 'features/SearchForm';
 import { reducer as sortMoviesReducer } from 'features/SearchMovie';
-import { reducer as movieDetailsReducer } from 'features/MovieDetails';
+import { reducer as movieDetailsReducer } from 'features/MovieCard';
 import { reducer as allMoviesReducer } from 'features/MoviesList';
 
 const formatParams = combineReducers({ searchBy: searchMoviesReducer, sortBy: sortMoviesReducer });
