+++ client/src/features/MovieCard/index.js
@@ -1,1 +1,2 @@
-export { default } from './component';
+export { default } from './container.enhanced';
+export { default as reducer } from './reducer';
