+++ client/src/features/MovieDetails/index.js
@@ -1,2 +1,1 @@
-export { default } from './container.enhanced';
-export { default as reducer } from './reducer';
+export { default } from './component';
